Fix nested boolean expressions
Nested boolean expressions were being dropped from the AST.

Also test NOT with ()s as the precedence rules make it apply to the
first token after the NOT, so `NOT foo == 5` means `(NOT foo) == 5`

Probably need to fix that someday, but for now we can just enforce the
use of ()s with NOTs.

Also started adding easier-to-read indentation to the tree parser.